# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto Backend
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  contact:
    email: deodati.2025953@studenti.uniroma1.it
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /feed:
    get:
      tags:
        - Feed tools
      summary: Retrieve feed
      description: Retrieves your feed made of the posts of the accounts you follow in reverse chronological order
      operationId: getMyStream
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '401':
          description: Unauthenticated
        '403':
          description: Trying to view somebody else's feed
...
  /users/{userID}/posts/{postID}/like:
    get:
      tags:
        - Feed tools
      summary: Get liked
      description: Checks whether the post has been liked by current user
      operationId: isLiked
      parameters:
        - in: path
          name: userID
          description: author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Operation successful
          content:
            text/plain:
              schema:
                type: boolean
        '401':
          description: Unauthenticated or bad authentication
        '404':
          description: Post not found
    put:
      tags:
        - Feed tools
      summary: Like a post
      description: Adds a like as the logged user. On success returns the new like count
      operationId: likePhoto
      parameters:
        - in: path
          name: userID
          description: author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '201':
          description: Operation successful
          content:
            text/plain:
              schema:
                type: integer
        '400':
          description: Bad auth token or bad post ID
        '401':
          description: Unauthenticated
        '403':
          description: "Forbidden: user blocked you!"
        '404':
          description: Post not found
    delete:
      tags:
        - Feed tools
      summary: Unlikes a post
      description: Removes the like as the logged user (if any). On success returns the new like count
      operationId: unlikePhoto
      parameters:
        - in: path
          name: userID
          description: author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Operation successful
          content:
            text/plain:
              schema:
                type: integer
        '400':
          description: Bad auth token, bad post ID or post wasn't liked
        '401':
          description: Unauthenticated
        '404':
          description: Post not found
  /users/{userID}/posts/{postID}/comments:
    post:
      tags:
        - Feed tools
      summary: Comment on a post
      description: Comments on a post. On success returns the new comment ID
      operationId: commentPhoto
      requestBody:
        content:
          text/plain:
            schema:
              $ref: '#/components/schemas/LargeText'
      parameters:
        - in: path
          name: userID
          description: post author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '201':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
        '400':
          description: Empty comment or comment too long
        '401':
          description: Unauthenticated
        '403':
          description: "Cannot comment no post: poster blocked you!"
        '404':
          description: Post not found
  /users/{userID}/posts/{postID}/comments/{commentID}:
    get:
      tags:
        - Feed tools
      summary: Get comment
      description: Retrieves the comment associated with the supplied ID
      operationId: getComment
      parameters:
        - in: path
          name: userID
          description: post author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: commentID
          description: Comment ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthenticated
        '403':
          description: "Forbidden: author blocked you!"
        '404':
          description: Comment not found
    delete:
      tags:
        - Feed tools
      summary: Remove a comment
      description: Removes the specified comment. Only the author can do that!
      operationId: uncommentPhoto
      parameters:
        - in: path
          name: userID
          description: post author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: commentID
          description: Comment ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Bad comment ID
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden, you cannot delete somebody else's comment!
        '404':
          description: Comment not found
  /users/{userID}/posts/{postID}/comments/{commentID}/like:
    get:
      tags:
        - Feed tools
      summary: Is comment liked?
      description: Checks whether the comment has been liked by you
      operationId: isCommentLiked
      parameters:
        - in: path
          name: userID
          description: post author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: commentID
          description: Comment ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Operation successful
          content:
            text/plain:
              schema:
                type: boolean
        '401':
          description: Unauthenticated or bad authentication
        '404':
          description: Post not found
    put:
      tags:
        - Feed tools
      summary: Likes a comment
      description: Adds a like to a comment as the logged user. On success returns the new like count
      operationId: likeComment
      parameters:
        - in: path
          name: userID
          description: post author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: commentID
          description: Comment ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '201':
          description: Operation successful
          content:
            text/plain:
              schema:
                type: integer
        '400':
          description: Bad authentication token or bad comment ID
        '401':
          description: Unauthenticated
        '403':
          description: "Cannot like: author blocked you!"
        '404':
          description: Comment not found
    delete:
      tags:
        - Feed tools
      summary: Unlike a comment
      description: Removes the like from a comment as the logged user (if any). On success returns the new like count
      operationId: unlikeComment
      parameters:
        - in: path
          name: userID
          description: post author ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: postID
          description: Post ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - in: path
          name: commentID
          description: Comment ID
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Operation successful
          content:
            text/plain:
              schema:
                type: integer
        '400':
          description: Bad authentication token, userID or comment ID
        '401':
          description: Unauthenticated
        '404':
          description: Comment not found or comment wasn't liked
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ID:
      type: integer
      format: uint64
      description: ID represents an integer used as an identifier (such as postID and commentID)
      minimum: 0
      example: 32156
    Counter:
      type: integer
      format: uint64
      description: A counter used to represent likes/comments/followers counts
      minimum: 0
      example: 80
...
    LargeText:
      type: string
      description: A large text, used in post captions and comments
      pattern: '^.*?$'
      minLength: 0
      maxLength: 2048
      example: Currently working on WASAPhoto!
...
    Comment:
      type: object
      description: An object representing a comment
      properties:
        id:
          $ref: '#/components/schemas/ID'
        author:
          $ref: '#/components/schemas/ID'
        likes:
          $ref: '#/components/schemas/Counter'
        content:
          $ref: '#/components/schemas/LargeText'
...
    Feed:
      type: array
      description: |-
        A list of posts (as postIDs) in reverse chronological order, representing a user's feed
      items:
        $ref: '#/components/schemas/ID'
security:
  - bearerAuth: []